LayoutTests/storage/filesystemaccess/writable-file-stream-seek-expected.txt

This test checks seek command of FileSystemWritableFileStream

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Run writer tests:
PASS fileContent is "abcdihgfe\u0000abcde"
Run stream tests:
PASS fileContent is "abcdihgfe\u0000abcde"
PASS successfullyParsed is true

TEST COMPLETE
